+++ src/cascader/types.ts
@@ -1,11 +1,15 @@
-export interface CascaderOption {
+export interface ThyCascaderOption {
     value?: any;
     label?: string;
     title?: string;
     disabled?: boolean;
     loading?: boolean;
     isLeaf?: boolean;
-    parent?: CascaderOption;
-    children?: CascaderOption[];
+    parent?: ThyCascaderOption;
+    children?: ThyCascaderOption[];
     [key: string]: any;
 }
+
+export type ThyCascaderTriggerType = 'click' | 'hover';
+
+export type ThyCascaderExpandTrigger = 'click' | 'hover';