+++ src/guider/utils.ts
@@ -1,14 +1,11 @@
-import { isArray, isNumber } from 'ngx-tethys/util';
+import { SafeAny } from 'ngx-tethys/types';
+import { isArray, isNumber, isObject } from 'ngx-tethys/util';
 
-export function isPositionDataType(target: any): target is [number, number] {
-    if (isArray(target)) {
-        if (target.length !== 2) {
-            return false;
-        }
-        if (isNumber(target[0]) && isNumber(target[1])) {
+export function isPositionDataType(target: any): target is { x: number; y: number } {
+    if (isObject(target) && !isArray(target)) {
+        if (isNumber((target as SafeAny).x) && isNumber((target as SafeAny).y)) {
             return true;
         }
-        return false;
     }
     return false;
 }