+++ src/shared/directives/thy-autofocus.directive.ts
@@ -2,6 +2,10 @@ import { Directive, Input, ElementRef, NgZone } from '@angular/core';
 import { reqAnimFrame } from 'ngx-tethys/core';
 import { coerceBooleanProperty } from 'ngx-tethys/util';
 
+/**
+ * 自动聚焦指令
+ * @name input[thyAutofocus],textarea[thyAutofocus]
+ */
 @Directive({
     selector: 'input[thyAutofocus],textarea[thyAutofocus]'
 })
@@ -9,12 +13,9 @@ export class ThyAutofocusDirective {
     // 自动选择，用于只读的 input 输入框，方便复制粘贴
     private _autoSelect = false;
 
-    @Input()
-    set thyAutoSelect(value: boolean) {
-        this._autoSelect = coerceBooleanProperty(value);
-    }
-
-    // auto focus current element
+    /**
+     * 是否自动聚焦
+     */
     @Input()
     set thyAutofocus(value: boolean | string) {
         if (coerceBooleanProperty(value) !== false) {
@@ -33,6 +34,14 @@ export class ThyAutofocusDirective {
                 })
             );
         }
+    }
+
+    /**
+     * 是否自动选择
+     */
+    @Input()
+    set thyAutoSelect(value: boolean) {
+        this._autoSelect = coerceBooleanProperty(value);
     }
 
     constructor(private elementRef: ElementRef, private ngZone: NgZone) {}