+++ src/shared/option/option-group.component.ts
@@ -2,6 +2,9 @@ import { Component, ViewEncapsulation, ChangeDetectionStrategy } from '@angular/
 import { ThyCanDisable, mixinDisabled, ThyCanDisableCtor, MixinBase } from 'ngx-tethys/core';
 
 const _MixinBase: ThyCanDisableCtor & typeof MixinBase = mixinDisabled(MixinBase);
+/**
+ * @private
+ */
 @Component({
     selector: 'thy-list-option-group,[thy-list-option-group]',
     template: '<ng-content></ng-content>',