+++ src/table/table.module.ts
@@ -4,6 +4,7 @@ import { ThyLoadingModule } from 'ngx-tethys/loading';
 import { ThyPaginationModule } from 'ngx-tethys/pagination';
 import { ThySharedModule } from 'ngx-tethys/shared';
 import { ThySwitchModule } from 'ngx-tethys/switch';
+import { ThySkeletonModule } from 'ngx-tethys/skeleton';
 
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { ScrollingModule } from '@angular/cdk/scrolling';
@@ -15,9 +16,16 @@ import { TableRowDragDisabledPipe } from './pipes/drag.pipe';
 import { TableIsValidModelValuePipe } from './pipes/table.pipe';
 import { ThyTableColumnComponent } from './table-column.component';
 import { ThyTableComponent } from './table.component';
+import { ThyTableSkeletonComponent } from './table-skeleton.component';
 
 @NgModule({
-    declarations: [ThyTableComponent, ThyTableColumnComponent, TableIsValidModelValuePipe, TableRowDragDisabledPipe],
+    declarations: [
+        ThyTableComponent,
+        ThyTableColumnComponent,
+        TableIsValidModelValuePipe,
+        TableRowDragDisabledPipe,
+        ThyTableSkeletonComponent
+    ],
     imports: [
         CommonModule,
         FormsModule,
@@ -26,10 +34,11 @@ import { ThyTableComponent } from './table.component';
         ThyLoadingModule,
         ThyEmptyModule,
         ThySharedModule,
+        ThySkeletonModule,
         ThyIconModule,
         DragDropModule,
         ScrollingModule
     ],
-    exports: [ThyTableComponent, ThyTableColumnComponent]
+    exports: [ThyTableComponent, ThyTableColumnComponent, ThyTableSkeletonComponent]
 })
 export class ThyTableModule {}